feat: Add Profile Model Admin
Created the ProfileAdmin class to customize the administration interface for the Profile model. This admin configuration defines fieldsets, list display options, autocomplete fields, ordering, search fields, and read-only fields for better management of user profiles in the Django admin panel.

feat: Add Signal for Profile Creation on User Creation

Implemented a signal handler to create a profile instance automatically whenever a new user instance is created. The signal is connected to the post_save signal of the custom user model. Additionally, registered the signal handler in the ready method of the app's apps.py file to ensure it is loaded when the application starts.